renaming test-animation-android to test-stack-animation-android
