+++ apps/src/tests/single-feature-tests/stack-v5/index.ts
@@ -4,7 +4,7 @@ import type { ScenarioGroup } from '@apps/tests/shared/helpers';
 // scenario group consumed by the selection menu.
 import TestStackPreventNativeDismissSingleStack from './prevent-native-dismiss-single-stack';
 import TestStackPreventNativeDismissNestedStack from './prevent-native-dismiss-nested-stack';
-import TestStackAnimationAndroid from './test-animation-android';
+import TestStackAnimationAndroid from './test-stack-animation-android';
 import TestStackSimpleNav from './test-stack-simple-nav';
 import TestStackSubviewsAndroid from './test-stack-subviews-android';
 import TestStackSubviewsIOS from './test-stack-subviews-ios';
@@ -24,7 +24,7 @@ import TestStackHeaderSelectiveUpdates from './test-stack-header-selective-updat
 // under a name for direct rendering (e.g. from App.tsx or e2e harnesses).
 export { default as TestStackPreventNativeDismissSingleStack } from './prevent-native-dismiss-single-stack';
 export { default as TestStackPreventNativeDismissNestedStack } from './prevent-native-dismiss-nested-stack';
-export { default as TestStackAnimationAndroid } from './test-animation-android';
+export { default as TestStackAnimationAndroid } from './test-stack-animation-android';
 export { default as TestStackSimpleNav } from './test-stack-simple-nav';
 export { default as TestStackSubviewsAndroid } from './test-stack-subviews-android';
 export { default as TestStackSubviewsIOS } from './test-stack-subviews-ios';
