+++ apps/src/tests/single-feature-tests/stack-v5/test-stack-animation-android/scenario-description.ts
